Attempt to update staging google sheet
