--- google-apps-script/src/example.ts
@@ -1,3 +1,0 @@
-export function helloWorld() {
-  console.log("Hello, world!");
-}
